@@ -1,5 +1,5 @@
 [build-system]
-requires = ['pip>=20.3', 'setuptools>=62', 'wheel']
+requires = ['pip>=24.2', 'setuptools>=74', 'wheel']
 build-backend = "setuptools.build_meta"
 
 [project]
@@ -8,26 +8,29 @@ authors = [{ name = "ForgQi" }]
 description = "stream download and upload"
 keywords = ["bilibili douyu huya"]
 urls = { Homepage = "https://github.com/ForgQi/bilibiliupload" }
-requires-python = ">= 3.7"
+requires-python = ">= 3.9"
 dependencies = [
     "tomli-w>=1.0.0",
-    "stream-gears >= 0.1.19,<0.2",
+    "stream-gears >= 0.2.2,<0.3",
     "psutil >= 5.4.6",
     "aiofiles >= 22.1.0",
-    "yt-dlp >= 2023.11.16",
+    "yt-dlp >= 2025.2.19",
     "Pillow >= 5.2.0",
-    "aiohttp[speedups]",
-    "aiohttp-cors",
-    "Requests >= 2.20.0",
+    "aiohttp[speedups] < 3.10.0", # aio-libs/aiohttp#10519
+    "aiohttp_cors >= 0.7.0",
+    "requests >= 2.32.3",
+    "httpx[http2] >= 0.28.1",
+    "httpcore >= 1.0.7",
     "PyYAML >= 4.2b1",
-    "streamlink >= 6.5.0",
-    "ykdl == 1.8.0; python_version < '3.8'",
-    "ykdl >= 1.8.0; python_version >= '3.8'",
+    "streamlink >= 7.1.3",
+    "ykdl >= 1.8.0",
     "rsa >= 4.6",
-    'importlib-resources; python_version<"3.9"',
-    'tomli >= 1.1.0 ; python_version < "3.11"',
-    'protobuf >= 4.23.0',
-    "peewee>=3.16.2",
+    "tomli >= 1.1.0; python_version < '3.11'",
+    "protobuf >= 4.23.0",
+    "SQLAlchemy >= 2.0.0",
+    "alembic >= 1.12.1",
+    "m3u8 >= 6.0.0",
+    "truststore >= 0.10.1; python_version >= '3.10'",
 ]
 dynamic = ["version"]
 
@@ -46,7 +49,7 @@ license-files = ["LICENSE"]
 include-package-data = false
 
 [tool.setuptools.package-data]
-biliup = ["common/*.ini", "web/public/**", "web/public/build/*"]
+biliup = ["common/*.ini", "web/public/**", "web/public/build/*", "database/migration/*", "Danmaku/douyin_util/webmssdk.js"]
 
 [tool.setuptools.packages.find]
 include = ["biliup*"]
