@@ -2,7 +2,7 @@ import logging
 import platform
 import sys
 
-__version__ = "0.4.34"
+__version__ = "0.4.92"
 
 LOG_CONF = {
     'version': 1,
@@ -30,7 +30,8 @@ LOG_CONF = {
             'interval': 1,
             'backupCount': 1,
             'filename': 'ds_update.log',
-            'formatter': 'verbose'
+            'formatter': 'verbose',
+            'encoding': 'utf-8'
         }
     },
     'root': {
@@ -52,3 +53,9 @@ if (3, 10, 6) > sys.version_info >= (3, 8) and platform.system() == 'Windows':
 
     proactor_events._ProactorBasePipeTransport.__del__ = silence_event_loop_closed(
         proactor_events._ProactorBasePipeTransport.__del__)
+
+
+if getattr(sys, 'frozen', False) and hasattr(sys, '_MEIPASS'):
+    import multiprocessing
+    multiprocessing.freeze_support()
+    print('running in a PyInstaller bundle')
