@@ -5,18 +5,13 @@ import asyncio
 import logging.config
 import platform
 import shutil
-import threading
-import time
-from pathlib import Path
 
 import biliup.common.reload
-from biliup.common.timer import Timer
 from biliup.config import config
-from biliup.database import DB as db
-from biliup.database.models import LiveStreamers, UploadStreamers
-from . import __version__, LOG_CONF
-from .common.Daemon import Daemon
-from .common.reload import AutoReload
+from biliup import __version__, LOG_CONF
+from biliup.common.Daemon import Daemon
+from biliup.common.reload import AutoReload
+from biliup.common.log import DebugLevelFilter
 
 
 def arg_parser():
@@ -31,6 +26,7 @@ def arg_parser():
     parser.add_argument('-v', '--verbose', action="store_const", const=logging.DEBUG, help="Increase output verbosity")
     parser.add_argument('--config', type=argparse.FileType(mode='rb'),
                         help='Location of the configuration file (default "./config.yaml")')
+    parser.add_argument('--no-access-log', action='store_true', help='disable web access log')
     subparsers = parser.add_subparsers(help='Windows does not support this sub-command.')
     # create the parser for the "start" command
     parser_start = subparsers.add_parser('start', help='Run as a daemon process.')
@@ -42,41 +38,63 @@ def arg_parser():
     parser.set_defaults(func=lambda: asyncio.run(main(args)))
     args = parser.parse_args()
     biliup.common.reload.program_args = args.__dict__
-    # 初始化数据库
-    if db.init():
-        try:
-            config.load(args.config)
-            config.save_to_db()
-        except FileNotFoundError:
-            print(f'新版本不依赖配置文件,请访问 http://ip:{args.port} 修改配置')
-    config.load_from_db()
-    LOG_CONF.update(config.get('LOGGING', {}))
-    if args.verbose:
-        LOG_CONF['loggers']['biliup']['level'] = args.verbose
-        LOG_CONF['root']['level'] = args.verbose
-    logging.config.dictConfig(LOG_CONF)
+
+    is_stop = args.func == daemon.stop
+
+    if not is_stop:
+        from biliup.database.db import SessionLocal, init
+        with SessionLocal() as db:
+            from_config = False
+            try:
+                config.load(args.config)
+                from_config = True
+            except FileNotFoundError:
+                print(f'新版本不依赖配置文件，请访问 WebUI 修改配置')
+            if init(args.no_http, from_config):
+                if from_config:
+                    config.save_to_db(db)
+            config.load_from_db(db)
+        # db.remove()
+        LOG_CONF.update(config.get('LOGGING', {}))
+        if args.verbose:
+            LOG_CONF['loggers']['biliup']['level'] = args.verbose
+            LOG_CONF['root']['level'] = args.verbose
+        logging.config.dictConfig(LOG_CONF)
+        logging.getLogger('httpx').addFilter(DebugLevelFilter())
+        # logging.getLogger('hpack').setLevel(logging.CRITICAL)
+        # logging.getLogger('httpx').setLevel(logging.CRITICAL)
     if platform.system() == 'Windows':
+        if is_stop:
+            return
         return asyncio.run(main(args))
     args.func()
 
 
 async def main(args):
-    from .app import event_manager, context
-    from biliup.downloader import check_flag
-    import biliup.web
+    from biliup.app import event_manager
 
     event_manager.start()
-    wait = max(config.get('event_loop_interval', 40) - 3, 3)
 
     # 启动时删除临时文件夹
     shutil.rmtree('./cache/temp', ignore_errors=True)
 
     interval = config.get('check_sourcecode', 15)
 
-    runner, site = await biliup.web.service(args)
-    detector = AutoReload(event_manager, runner.cleanup, check_flag.set, interval=interval)
-    biliup.common.reload.global_reloader = detector
-    await asyncio.gather(detector.astart(), site.start())
+    if not args.no_http:
+        import biliup.web
+        runner = await biliup.web.service(args)
+        detector = AutoReload(event_manager, runner.cleanup, interval=interval)
+        biliup.common.reload.global_reloader = detector
+        await detector.astart()
+    else:
+        import biliup.common.reload
+        detector = AutoReload(event_manager, interval=interval)
+        biliup.common.reload.global_reloader = detector
+        await asyncio.gather(detector.astart())
+
+
+class GracefulExit(SystemExit):
+    code = 1
 
 
 if __name__ == '__main__':
