@@ -1,11 +1,10 @@
 import json
 import pathlib
 import shutil
+import os
 from collections import UserDict
+from sqlalchemy import select
 
-from playhouse.shortcuts import model_to_dict
-
-from biliup.database.models import Configuration, LiveStreamers, db, UploadStreamers
 
 try:
     import tomllib
@@ -15,7 +14,13 @@ except ModuleNotFoundError:
 
 class Config(UserDict):
     def load_cookies(self, file='cookies.json'):
-        self.data["user"] = {"cookies": {}}
+        if not os.path.exists(file):
+            raise FileNotFoundError(file)
+
+        if self.data.get('user') is None:
+            self.data['user'] = {'cookies': {}}
+        if self.data["user"].get("cookies") is None:
+            self.data['user']['cookies'] = {}
         with open(file, encoding='utf-8') as stream:
             s = json.load(stream)
             for i in s["cookie_info"]["cookies"]:
@@ -23,36 +28,50 @@ class Config(UserDict):
                 self.data["user"]["cookies"][name] = i["value"]
             self.data["user"]["access_token"] = s["token_info"]["access_token"]
 
-    def load_from_db(self):
+    def load_from_db(self, db):
+        from biliup.database.models import Configuration, LiveStreamers
         context = {
             'url_upload_count': self.data.get('url_upload_count', {}),
             'upload_filename': self.data.get('upload_filename', []),
-            'PluginInfo': self.data.get('PluginInfo')
+            'PluginInfo': self.data.get('PluginInfo'),
+            'sync_downloader_map': self.data.get('sync_downloader_map', {}),
         }
-        for con in Configuration.select().where(Configuration.key == 'config'):
+
+        for con in db.execute(select(Configuration.value).where(Configuration.key == 'config')):
             self.data = json.loads(con.value)
         self.data.update(context)
         self['streamers'] = {}
-        for ls in LiveStreamers.select():
-            self['streamers'][ls.remark] = {k: v for k, v in model_to_dict(ls).items() if v and (k != 'upload_streamers')}
+        for ls in db.scalars(select(LiveStreamers)):
+            self['streamers'][ls.remark] = {
+                k: v for k, v in ls.as_dict().items() if v and (k not in ['upload_streamers_id', 'id', 'remark'])}
             # self['streamers'][ls.remark].pop('upload_streamers')
-            if ls.upload_streamers:
-                self['streamers'][ls.remark].update({k: v for k, v in model_to_dict(ls.upload_streamers).items() if v})
-            if self['streamers'][ls.remark].get('tags'):
-                self['streamers'][ls.remark]['tags'] = self['streamers'][ls.remark]['tags']
+            if ls.upload_streamers_id:
+                self['streamers'][ls.remark].update({
+                    k: v for k, v in ls.uploadstreamers.as_dict().items() if v and k not in ['id', 'template_name']})
+                if self['streamers'][ls.remark].get('uploader') is None:
+                    self['streamers'][ls.remark]['uploader'] = 'biliup-rs'
+            # if self['streamers'][ls.remark].get('tags'):
+            #     self['streamers'][ls.remark]['tags'] = self['streamers'][ls.remark]['tags']
         # for us in UploadStreamers.select():
         #     config.data[con.key] = con.value
 
-    def save_to_db(self):
-        with db.connection_context():
-            for k, v in self['streamers'].items():
-                us = UploadStreamers(template_name=k, tags=v.pop('tags', ['biliup']), **v)
-                us.save()
-                for url in v.pop('url'):
-                    LiveStreamers(upload_streamers=us, remark=k, url=url, **v).save()
-            del self['streamers']
-            Configuration(key='config', value=json.dumps(self.data)).save()
-
+    def save_to_db(self, db):
+        from biliup.database.models import Configuration, LiveStreamers, UploadStreamers
+        for k, v in self['streamers'].items():
+            us = UploadStreamers(**UploadStreamers.filter_parameters(
+                {"template_name": k, "tags": v.pop('tags', [k]), ** v}))
+            db.add(us)
+            db.flush()
+            url = v.pop('url')
+            urls = url if isinstance(url, list) else [url]  # 兼容 url 输入字符串和列表
+            for url in urls:
+                ls = LiveStreamers(**LiveStreamers.filter_parameters(
+                    {"upload_streamers_id": us.id, "remark": k, "url": url, ** v}))
+                db.add(ls)
+        del self['streamers']
+        configuration = Configuration(key='config', value=json.dumps(self.data))
+        db.add(configuration)
+        db.commit()
 
     def load(self, file):
         import yaml
@@ -121,6 +140,30 @@ class Config(UserDict):
                 old_data["threads"] = self.data["threads"]
                 old_data["streamers"] = self.data["streamers"]
                 yaml.dump(old_data, stream, default_flow_style=False, allow_unicode=True)
+
+    def dump(self, file):
+        """ 转储到旧版配置文件 """
+        if not file:
+            file = 'config.toml'
+        if os.path.exists(file):
+            from datetime import datetime
+            import logging
+            logger = logging.getLogger('biliup')
+            new_name = f'{file}.backup.{datetime.now().strftime("%Y%m%d%H%M%S")}'
+            logger.info(f"{file} 文件已存在，已将原文件重命名为 {new_name}")
+            os.rename(file, new_name)
+        exclude_keys = ['PluginInfo', 'upload_filename', 'url_upload_count']
+        temp = {k: v for k, v in self.data.items() if k not in exclude_keys}
+        if self.data.get('yaml') or file.endswith(".yaml"):
+            import yaml
+            with open(file, 'w+', encoding='utf-8') as stream:
+                yaml.dump(temp, stream, default_flow_style=False, allow_unicode=True)
+        else:
+            import tomli_w
+            # print(config)
+            with open(file, 'wb') as stream:
+                tomli_w.dump(temp, stream)
+        return file
 
 
 config = Config()
