@@ -1,4 +1,3 @@
-import copy
 import json
 import logging
 import os
@@ -11,14 +10,14 @@ from typing import List
 
 from biliup.config import config
 from .app import event_manager, context
-from .common.tools import NamedLock
-from .downloader import download, send_upload_event
+from .common.tools import NamedLock, processor
+from .common.util import check_timerange
+from .database.db import get_stream_info_by_filename, SessionLocal
+from .downloader import biliup_download
 from .engine.event import Event
 from .engine.upload import UploadBase
 from .uploader import upload, fmt_title_and_desc
-from biliup.database import DB
 
-CHECK = 'check'
 PRE_DOWNLOAD = 'pre_download'
 DOWNLOAD = 'download'
 DOWNLOADED = 'downloaded'
@@ -27,34 +26,17 @@ UPLOADED = 'uploaded'
 logger = logging.getLogger('biliup')
 
 
-@event_manager.register(CHECK, block='Asynchronous1')
-def singleton_check(platform, name, url):
-    from .plugins.twitch import Twitch
-    if platform == Twitch:
-        # 如果支持批量检测，目前只有一个支持，第一版先写死按照特例处理
-        for turl in Twitch.batch_check.__func__(Twitch.url_list):
-            yield Event(PRE_DOWNLOAD, args=(name, turl,))
-        return
-    if context['PluginInfo'].url_status[url] == 1:
-        logger.info(f'{url}-{url}-正在下载中，跳过检测')
-        return
-    # 可能对同一个url同时发送两次上传事件
-    with NamedLock(f"upload_count_{url}"):
-        # from .handler import event_manager, UPLOAD
-        # += 不是原子操作
-        context['url_upload_count'].setdefault(url, 0)
-        context['url_upload_count'][url] += 1
-        yield Event(UPLOAD, ({'name': name, 'url': url},))
-    if platform(name, url).check_stream(True):
-        # 需要等待上传文件列表检索完成后才可以开始下次下载
-        with NamedLock(f'upload_file_list_{name}'):
-            context['PluginInfo'].url_status[url] = 1
-            yield Event(PRE_DOWNLOAD, args=(name, url,))
+# @event_manager.register(CHECK, block='Asynchronous3')
 
 
 @event_manager.register(PRE_DOWNLOAD, block='Asynchronous1')
 def pre_processor(name, url):
-    logger.info(f'{name}-{url}-开播了准备下载')
+    url_status = context["PluginInfo"].url_status
+    if url_status[url] == 1:
+        logger.debug(f"{name} - {url} 正在下载中，跳过下载")
+        return
+
+    logger.debug(f"{name} - {url} 开播了准备下载")
     preprocessor = config['streamers'].get(name, {}).get('preprocessor')
     if preprocessor:
         processor(preprocessor, json.dumps({
@@ -64,36 +46,26 @@ def pre_processor(name, url):
         }, ensure_ascii=False))
     yield Event(DOWNLOAD, (name, url))
 
+
 @event_manager.register(DOWNLOAD, block='Asynchronous1')
 def process(name, url):
-    stream_info = {
-        'name': name,
-        'url': url,
-    }
     url_status = context['PluginInfo'].url_status
     # 下载开始
     try:
-        kwargs: dict = config['streamers'][name].copy()
-        kwargs.pop('url')
-        suffix = kwargs.get('format')
-        if suffix:
-            kwargs['suffix'] = suffix
-        stream_info = download(name, url, **kwargs)
+        url_status[url] = 1
+        stream_info = biliup_download(name, url, config['streamers'][name].copy())
+        yield Event(DOWNLOADED, (stream_info,))
     except Exception as e:
-        logger.exception(f"下载错误: {stream_info['name']} - {e}")
+        logger.exception(f"下载错误: {name} - {e}")
     finally:
         # 下载结束
-        # 永远不可能有两个同url的下载线程
-        # 可能对同一个url同时发送两次上传事件
-        with NamedLock(f"upload_count_{stream_info['url']}"):
-            # += 不是原子操作
-            context['url_upload_count'][stream_info['url']] += 1
-            yield Event(DOWNLOADED, (stream_info,))
-        url_status[url] = 0
+        url_status[url] = 0 if check_timerange(name) else 2
+
 
 @event_manager.register(DOWNLOADED, block='Asynchronous1')
 def processed(stream_info):
     name = stream_info['name']
+    url = stream_info['url']
     # 下载后处理 上传前处理
     downloaded_processor = config['streamers'].get(name, {}).get('downloaded_processor')
     if downloaded_processor:
@@ -101,7 +73,7 @@ def processed(stream_info):
         file_list = UploadBase.file_list(name)
         processor(downloaded_processor, json.dumps({
             "name": name,
-            "url": stream_info.get('url'),
+            "url": url,
             "room_title": stream_info.get('title', name),
             "start_time": int(time.mktime(stream_info.get('date', default_date))),
             "end_time": int(time.mktime(stream_info.get('end_time', default_date))),
@@ -116,20 +88,47 @@ def process_upload(stream_info):
     url = stream_info['url']
     name = stream_info['name']
     url_upload_count = context['url_upload_count']
+    # 永远不可能有两个同url的下载线程
+    # 可能对同一个url同时发送两次上传事件
+    with NamedLock(f"upload_count_{url}"):
+        if context['url_upload_count'][url] > 0:
+            return logger.debug(f'{url} 正在上传中，跳过')
+        # from .handler import event_manager, UPLOAD
+        # += 不是原子操作
+        context['url_upload_count'][url] += 1
     # 上传开始
     try:
         file_list = UploadBase.file_list(name)
-        if len(file_list) <= 0:
+        file_count = len(file_list)
+        if file_count <= 0:
             logger.debug("无需上传")
             return
+
+        # 上传延迟检测
+        url_status = context['PluginInfo'].url_status
+        if url_status[url] != 2:
+            delay = int(config.get('delay', 0))
+            if delay:
+                logger.info(f'{name} -> {url} {delay}s 后检测是否上传')
+                time.sleep(delay)
+                if url_status[url] == 1:
+                    # 上传延迟检测，启用的话会在一段时间后检测是否存在下载任务，若存在则跳过本次上传
+                    return logger.info(f'{name} -> {url} 存在下载任务, 跳过本次上传')
+
         if ("title" not in stream_info) or (not stream_info["title"]):  # 如果 data 中不存在标题, 说明下载信息已丢失, 则尝试从数据库获取
-            data, _ = fmt_title_and_desc({
-                **DB.get_stream_info_by_filename(os.path.splitext(file_list[0].video)[0]),
-                "name": name})  # 如果 restart, data 中会缺失 name 项
+            with SessionLocal() as db:
+                i = 0
+                # 按创建时间排序可能导致首个结果无法获取到数据
+                while i < file_count:
+                    data = get_stream_info_by_filename(db, file_list[i].video)
+                    if data:
+                        break
+                    i += 1
+            data, _ = fmt_title_and_desc({**data, "name": name})  # 如果 restart, data 中会缺失 name 项
             stream_info.update(data)
         filelist = upload(stream_info)
         if filelist:
-            yield Event(UPLOADED, (name, stream_info.get('live_cover_path'), filelist))
+            uploaded(name, stream_info.get('live_cover_path'), filelist)
     except Exception:
         logger.exception(f"上传错误: {name}")
     finally:
@@ -138,7 +137,7 @@ def process_upload(stream_info):
         with NamedLock(f'upload_count_{url}'):
             url_upload_count[url] -= 1
 
-@event_manager.register(UPLOADED, block='Asynchronous2')
+
 def uploaded(name, live_cover_path, data: List):
     # data = file_list
     post_processor = config['streamers'].get(name, {}).get("postprocessor", None)
@@ -155,7 +154,15 @@ def uploaded(name, live_cover_path, data: List):
             file_list.append(i.danmaku)
 
     for post_processor in post_processor:
-        if post_processor == 'rm':
+        if  ( # 兼容新WEB字段
+                isinstance(post_processor, str)
+                and
+                post_processor == 'rm'
+            ) or (
+                isinstance(post_processor, dict)
+                and
+                post_processor.get('rm')
+            ):
             # 删除封面
             if live_cover_path is not None:
                 UploadBase.remove_file(live_cover_path)
@@ -178,20 +185,6 @@ def uploaded(name, live_cover_path, data: List):
                 process_output = subprocess.check_output(
                     post_processor['run'], shell=True,
                     input=reduce(lambda x, y: x + str(Path(y).absolute()) + '\n', file_list, ''),
-                    stderr=subprocess.STDOUT, text=True)
-                logger.info(process_output.rstrip())
-            except subprocess.CalledProcessError as e:
-                logger.exception(e.output)
-                continue
-
-
-def processor(processors, data):
-    for processor in processors:
-        if processor.get('run'):
-            try:
-                process_output = subprocess.check_output(
-                    processor['run'], shell=True,
-                    input=data,
                     stderr=subprocess.STDOUT, text=True)
                 logger.info(process_output.rstrip())
             except subprocess.CalledProcessError as e:
