@@ -2,7 +2,10 @@ import random
 import string
 import json
 import requests
+
+import biliup.common.util
 from . import logger
+from ..common import tools
 from ..engine.decorators import Plugin
 from ..engine.download import DownloadBase
 
@@ -12,7 +15,7 @@ class Acfun(DownloadBase):
     def __init__(self, fname, url, suffix='flv'):
         super().__init__(fname, url, suffix)
 
-    def check_stream(self, is_check=False):
+    async def acheck_stream(self, is_check=False):
         if len(self.url.split("acfun.cn/live/")) < 2:
             logger.debug("直播间地址错误")
             return False
@@ -20,8 +23,8 @@ class Acfun(DownloadBase):
         did = "web_"+get_random_name(16)
         cookies = dict(_did=did)
         data1 = {'sid': 'acfun.api.visitor'}
-        r1 = requests.post("https://id.app.acfun.cn/rest/app/visitor/login",
-                           headers=self.fake_headers, data=data1, cookies=cookies, timeout=5)
+        r1 = await biliup.common.util.client.post("https://id.app.acfun.cn/rest/app/visitor/login",
+                                                  headers=self.fake_headers, data=data1, cookies=cookies)
         userid = r1.json()['userId']
         visitorst = r1.json()['acfun.api.visitor_st']
         params = {
@@ -34,8 +37,8 @@ class Acfun(DownloadBase):
         }
         data2 = {'authorId': rid, 'pullStreamType': 'FLV'}
         self.fake_headers['referer'] = "https://live.acfun.cn/"
-        r2 = requests.post("https://api.kuaishouzt.com/rest/zt/live/web/startPlay",
-                           headers=self.fake_headers, data=data2, params=params, timeout=5)
+        r2 = await biliup.common.util.client.post("https://api.kuaishouzt.com/rest/zt/live/web/startPlay",
+                                                  headers=self.fake_headers, data=data2, params=params)
         if r2.json().get('result') != 1:
             logger.debug(r2.json())
             return False
