@@ -1,5 +1,5 @@
-import requests
-
+import biliup.common.util
+from ..common import tools
 from ..engine.decorators import Plugin
 from ..engine.download import DownloadBase
 from ..plugins import logger
@@ -10,16 +10,16 @@ class Bigo(DownloadBase):
     def __init__(self, fname, url, suffix='flv'):
         super().__init__(fname, url, suffix)
 
-    def check_stream(self, is_check=False):
+    async def acheck_stream(self, is_check=False):
         try:
             room_id = self.url.split('/')[-1].split('?')[0]
         except:
             logger.warning(f"{Bigo.__name__}: {self.url}: 直播间地址错误")
             return False
         try:
-            room_info = requests.post(f'https://ta.bigo.tv/official_website/studio/getInternalStudioInfo', timeout=10,
-                                      headers={**self.fake_headers, 'Accept': 'application/json'},
-                                      data={"siteId": room_id}).json()
+            room_info = (await biliup.common.util.client.post(f'https://ta.bigo.tv/official_website/studio/getInternalStudioInfo', timeout=10,
+                                                             headers={**self.fake_headers, 'Accept': 'application/json'},
+                                                             data={"siteId": room_id})).json()
             if room_info['code'] != 0:
                 raise
         except:
