@@ -13,12 +13,14 @@ class BiliWeb(UploadBase):
     def __init__(
             self, principal, data, submit_api=None, copyright=2, postprocessor=None, dtime=None,
             dynamic='', lines='AUTO', threads=3, tid=122, tags=None, cover_path=None, description='',
-            dolby=0, hires=0, no_reprint=0, open_elec=0, credits=[],
-            user_cookie='cookies.json'
+            dolby=0, hires=0, no_reprint=0, open_elec=0, credits=None,
+            user_cookie='cookies.json', copyright_source=None, extra_fields = ""
     ):
         super().__init__(principal, data, persistence_path='bili.cookie', postprocessor=postprocessor)
         if tags is None:
             tags = []
+        else:
+            tags = [str(tag).format(streamer=self.data['name']) for tag in tags]
         self.lines = lines
         self.submit_api = submit_api
         self.threads = threads
@@ -31,7 +33,7 @@ class BiliWeb(UploadBase):
         else:
             self.cover_path = None
         self.desc = description
-        self.credits = credits
+        self.credits = credits if credits else []
         self.dynamic = dynamic
         self.copyright = copyright
         self.dtime = dtime
@@ -40,7 +42,9 @@ class BiliWeb(UploadBase):
         self.no_reprint = no_reprint
         self.open_elec = open_elec
         self.user_cookie = user_cookie
-
+        self.copyright_source = copyright_source
+        self.extra_fields = extra_fields
+        
     def upload(self, file_list: List[UploadBase.FileInfo]) -> List[UploadBase.FileInfo]:
         if self.credits:
             desc_v2 = self.creditsToDesc_v2()
@@ -61,7 +65,7 @@ class BiliWeb(UploadBase):
             "tid": self.tid,
             "tag": ','.join(self.tags),
             "copyright": self.copyright,
-            "source": self.data["url"] if self.copyright == 2 else "",
+            "source": self.copyright_source if self.copyright_source else self.data["url"],
             "desc": self.desc,
             "dynamic": self.dynamic,
             "cover": self.cover_path if self.cover_path is not None else "",
@@ -71,9 +75,9 @@ class BiliWeb(UploadBase):
             "open_elec": self.open_elec,
             "limit": self.threads,
             "desc_v2": desc_v2,
+            "extra_fields": self.extra_fields,
             "dtime": int(time.time() + self.dtime) if self.dtime else None,
         }
-
         upload_process = mp.get_context('spawn').Process(target=stream_gears_upload, daemon=True, kwargs=upload_args)
         upload_process.start()
         upload_process.join()
@@ -115,18 +119,18 @@ class BiliWeb(UploadBase):
 
 def stream_gears_upload(ex_conn, lines, *args, **kwargs):
     try:
-        if lines == 'kodo':
-            kwargs['line'] = stream_gears.UploadLine.Kodo
+        if lines == 'bda':
+            kwargs['line'] = stream_gears.UploadLine.Bda
         elif lines == 'bda2':
             kwargs['line'] = stream_gears.UploadLine.Bda2
         elif lines == 'ws':
             kwargs['line'] = stream_gears.UploadLine.Ws
         elif lines == 'qn':
             kwargs['line'] = stream_gears.UploadLine.Qn
-        elif lines == 'cos':
-            kwargs['line'] = stream_gears.UploadLine.Cos
-        elif lines == 'cos-internal':
-            kwargs['line'] = stream_gears.UploadLine.CosInternal
+        elif lines == 'tx':
+            kwargs['line'] = stream_gears.UploadLine.Tx
+        elif lines == 'txa':
+            kwargs['line'] = stream_gears.UploadLine.Txa
         elif lines == 'bldsa':
             kwargs['line'] = stream_gears.UploadLine.Bldsa
 
