@@ -11,7 +11,7 @@ class YDownload(DownloadBase):
         super().__init__(fname, url, suffix)
         self.ydl_opts = {}
 
-    def check_stream(self, is_check=False):
+    async def acheck_stream(self, is_check=False):
         try:
             self.get_sinfo()
             return True
@@ -32,8 +32,9 @@ class YDownload(DownloadBase):
             logger.debug(info_list)
         return info_list
 
-    def download(self, filename):
+    def download(self):
         try:
+            filename = self.gen_download_filename(is_fmt=True) + '.' + self.suffix
             self.ydl_opts = {'outtmpl': filename}
             with yt_dlp.YoutubeDL(self.ydl_opts) as ydl:
                 ydl.download([self.url])
@@ -48,7 +49,7 @@ class SDownload(DownloadBase):
         self.stream = None
         self.flag = Event()
 
-    def check_stream(self, is_check=False):
+    async def acheck_stream(self, is_check=False):
         logger.debug(self.fname)
         import streamlink
         try:
@@ -62,8 +63,8 @@ class SDownload(DownloadBase):
         except streamlink.StreamlinkError:
             return
 
-    def download(self, filename):
-
+    def download(self):
+        filename = self.gen_download_filename(is_fmt=True) + '.' + self.suffix
         # fd = stream.open()
         try:
             with self.stream.open() as fd:
@@ -75,7 +76,7 @@ class SDownload(DownloadBase):
                             return 1
                     return 0
         except OSError:
-            self.rename(filename)
+            self.download_file_rename(filename + '.part', filename)
             raise
 
 
@@ -84,7 +85,7 @@ class Generic(DownloadBase):
         super().__init__(fname, url, suffix)
         self.handler = self
 
-    def check_stream(self, is_check=False):
+    async def acheck_stream(self, is_check=False):
         logger.debug(self.fname)
         try:
             site, url = url_to_module(self.url)
@@ -96,17 +97,17 @@ class Generic(DownloadBase):
         except:
             handlers = [YDownload(self.fname, self.url, 'mp4'), SDownload(self.fname, self.url, 'flv')]
             for handler in handlers:
-                if handler.check_stream():
+                if await handler.acheck_stream():
                     self.handler = handler
                     self.suffix = handler.suffix
                     return True
             return False
         return True
 
-    def download(self, filename):
+    def download(self):
         if self.handler == self:
-            return super(Generic, self).download(filename)
-        return self.handler.download(filename)
+            return super(Generic, self).download()
+        return self.handler.download()
 
 
 __plugin__ = Generic
