@@ -1,29 +1,27 @@
 import json
-import urllib.request
-
-import requests
 import re
+
+import biliup.common.util
 from . import logger
-from biliup.config import config
+from ..common import tools
 from ..engine.decorators import Plugin
 from ..engine.download import DownloadBase
-from biliup.plugins.Danmaku import DanmakuClient
 
 
 @Plugin.download(regexp=r'(?:https?://)?(?:(?:www)\.)?inke\.cn')
-class inke(DownloadBase):
+class Inke(DownloadBase):
     def __init__(self, fname, url, suffix='flv'):
         super().__init__(fname, url, suffix)
 
-    def check_stream(self, is_check=False):
+    async def acheck_stream(self, is_check=False):
         logger.debug(self.fname)
         rid = re.search(r'uid=([a-zA-Z0-9]+)', self.url).group(1)
-        r1 = requests.get(
+        r1 = await biliup.common.util.client.get(
             f"https://webapi.busi.inke.cn/web/live_share_pc?uid={rid}",
             timeout = 5,
             headers = self.fake_headers
         )
-        r1.close()
+        # r1.close()
         jsons = json.loads(r1.text)
         if jsons:
             if jsons.get('error_code') != 0:
